authserver/requirements.txt
aio-pika==5.6.2
aiormq==2.7.2
idna==2.8
multidict==4.5.2
pamqp==2.3.0
PyJWT==1.7.1
typing-extensions==3.7.4
yarl==1.3.0
